Refactor the local HDF file writer test case

+++ src/DemuxTopic.cxx
@@ -37,7 +37,7 @@ DemuxTopic::DemuxTopic(std::string topic)
 
 DemuxTopic::DT DemuxTopic::time_difference_from_message(char *msg_data,
                                                         int msg_size) {
-  Msg msg{msg_data, msg_size};
+  Msg msg{msg_data, size_t(msg_size)};
   std::string _tmp_dummy;
   auto reader = FBSchemaReader::create(msg);
   if (!reader) {
@@ -51,7 +51,7 @@ DemuxTopic::DT DemuxTopic::time_difference_from_message(char *msg_data,
 std::string const &DemuxTopic::topic() const { return _topic; }
 
 ProcessMessageResult DemuxTopic::process_message(char *msg_data, int msg_size) {
-  Msg msg{msg_data, msg_size};
+  Msg msg{msg_data, size_t(msg_size)};
   auto reader = FBSchemaReader::create(msg);
   if (!reader) {
     return ProcessMessageResult::ERR();
